data/home/hmtf000001/extract_0329M/extract_0329G/Live_Trading_System/api/binance_api.py
from binance.client import Client
import os

class BinanceClient:
    client = Client(os.getenv("BINANCE_API_KEY"), os.getenv("BINANCE_SECRET_KEY"))

    @staticmethod
    def get_price(symbol):
        return BinanceClient.client.get_symbol_ticker(symbol=symbol)